feat(public-cloud): add application bootstrap
