+++ packages/manager/apps/public-cloud/package.json
@@ -12,10 +12,7 @@
   "author": "OVH SAS",
   "main": "./src/index.js",
   "scripts": {
-    "build": "yarn build:eu && yarn build:ca && yarn build:us",
-    "build:ca": "webpack --env.production --env.region=\"CA\" && mv dist dist-CA",
-    "build:eu": "webpack --env.production && mv dist dist-EU",
-    "build:us": "webpack --env.production --env.region=\"US\" && mv dist dist-US",
+    "build": "webpack --env.production ",
     "dev": "webpack-dev-server",
     "dev:watch": "yarn run dev",
     "start": "lerna exec --stream --scope='@ovh-ux/manager-public-cloud' --include-dependencies -- npm run build --if-present",
@@ -29,6 +26,7 @@
     "@ovh-ux/manager-error-page": "^1.0.0",
     "@ovh-ux/manager-navbar": "^2.2.0",
     "@ovh-ux/manager-pci": "^3.0.0",
+    "@ovh-ux/manager-preloader": "^0.0.0 || ^1.0.0",
     "@ovh-ux/manager-telecom-styles": "^3.1.0",
     "@ovh-ux/mfa-enrollment": "^0.0.0 || ^1.0.0",
     "@ovh-ux/ng-at-internet": "^4.0.4",
@@ -79,7 +77,8 @@
     "ovh-ui-angular": "^3.15.1",
     "ovh-ui-kit": "^2.40.0",
     "ovh-ui-kit-bs": "^2.3.1",
-    "tweetnacl-util": "^0.15.0"
+    "tweetnacl-util": "^0.15.0",
+    "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
     "@ovh-ux/manager-webpack-config": "^3.0.8",
