--- packages/manager/apps/public-cloud/src/components/manager-preload/translations/Messages_de_DE.json
@@ -1,4 +1,0 @@
-{
- "preload_welcome": "Herzlich willkommen in Ihrem",
- "preload_manager_ovh": "OVH Kundencenter"
-}
