--- packages/manager/apps/public-cloud/src/components/manager-preload/translations/Messages_fr_FR.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Bienvenue sur votre",
-  "preload_manager_ovh": "Espace client OVHcloud"
-}
