--- packages/manager/apps/public-cloud/src/components/manager-preload/translations/Messages_it_IT.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Benvenuto",
-  "preload_manager_ovh": "Spazio Cliente OVHcloud"
-}
