--- packages/manager/apps/public-cloud/src/components/manager-preload/translations/Messages_lt_LT.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Welcome to the",
-  "preload_manager_ovh": "OVHcloud Control Panel"
-}
