--- packages/manager/apps/public-cloud/src/components/manager-preload/translations/Messages_pl_PL.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Witamy w",
-  "preload_manager_ovh": "Panel klienta OVHcloud"
-}
