--- packages/manager/apps/public-cloud/src/components/manager-preload/translations/Messages_pt_PT.json
@@ -1,4 +1,0 @@
-{
-  "preload_welcome": "Bem-vindo(a) à",
-  "preload_manager_ovh": "Área de Cliente OVHcloud"
-}
