+++ packages/manager/apps/public-cloud/src/index.html
@@ -13,8 +13,6 @@
     <body class="ovh-public-cloud-app h-100" style="overflow: hidden;">
         <div
             class="h-100 d-flex flex-column overflow-hidden"
-            data-ng-app="ovhPublicCloudApp"
-            data-ng-strict-di
             data-ng-controller="PublicCloudController as $ctrl"
         >
             <div>
@@ -30,28 +28,6 @@
                     style="top: 0; left: 0;"
                 >
                     <div class="h-100 w-100" data-ui-view></div>
-                </div>
-            </div>
-            <div
-                id="managerPreload"
-                class="d-flex text-center"
-                data-ng-class="managerPreloadHide"
-            >
-                <div class="m-auto">
-                    <p
-                        class="lead mb-3"
-                        data-ng-if="('preload_welcome' | translate) !== 'preload_welcome'"
-                        data-translate="preload_welcome"
-                    ></p>
-                    <hr class="my-0" />
-                    <h1
-                        class="mt-3 mb-5"
-                        data-ng-if="('preload_manager_ovh' | translate) !== 'preload_manager_ovh'"
-                        data-translate="preload_manager_ovh"
-                    ></h1>
-                    <div class="mt-5">
-                        <oui-spinner data-size="l"></oui-spinner>
-                    </div>
                 </div>
             </div>
